renaming module
